Fix FindStandardMathLibraryC.cmake reusing variable names
The original version of the FindStandardMathLibrary test has been copied from elsewhere, but that version has
been intended to check the need of linking the math lib from C++.
libxc has adapted this test to instead check whether it is needed from C.

The problem here is that during this change, the variable names for the result
of the performed compile tests have been left untouched.
This becomes a problem when building libxc alongside another project that has
copied the same original test but leaves that test to work for C++ (as e.g. Eigen does).
Since CHECK_C_SOURCE_COMPILES does not re-perform the test when its result
variable is already available from cache, this can lead to situations where another
version of FindStandardMathLibrary that works for C++ is performed first (almost
certainly indicating that no additional linker flag is required) and once it's libxc's 
turn, the cached result will be reused. This, however, is nonesense as the cached
result is valid only for C++, but libxc interprets it as being valid for C. Thus, this can
lead to omitting the `-lm` flag wrongfully.